test(media): split local-root tests into their own file
Keep tests/unit/media-utils.test.ts under the repo's 500-line target by
moving the mediaLocalRoots boundary tests to
media-utils-local-roots.test.ts, guard the symlink cases on Windows, and
make temp-dir cleanup exception-safe.
